src/requirements.txt
alembic==1.6.5
astroid==2.6.2
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
isort==5.9.1
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pylint==2.9.3
python-dateutil==2.8.1
python-editor==1.0.4
semantic-version==2.8.5
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
Werkzeug==2.0.1
wrapt==1.12.1
